refactor: NameInputView 에서 LadderGameOperatorInputView 분리
